Added assets and made it responsive
