+++ src/landing_v2/landing_data.tsx
@@ -1,0 +1,127 @@
+const prerequisite_data = [
+  {
+    name: 'Basic Understanding of English and Maths',
+    descp: null,
+  },
+  {
+    name: 'Basic Understanding of Python Language',
+    descp:
+      'If not Please do this course and learn the basics, should not take more than 10 days, Feel free to ask your doubts in the python-beginner challenge.',
+  },
+  {
+    name:
+      'Be a part of community, be ready to spend 2-3 hours a week with your peers learning together',
+    descp: null,
+  },
+  {
+    name: 'Motivation to learn',
+    descp: null,
+  },
+];
+
+const who_can_join_data = [
+  {
+    name: 'If you looking to explore a career in coding(Complete beginner)',
+  },
+  {
+    name: 'You are passionate about coding and building stuff',
+  },
+  {
+    name:
+      'You want to revise a few concepts and you love interacting with the community',
+  },
+];
+
+const mentorsData = [
+  {
+    name: 'Vedansh Garg',
+    company: 'Ex - HackerRank',
+    img: './mentor_ved.jpg',
+    linkedin: 'https://www.linkedin.com/in/vedanshgarg/',
+    descp: `Vedansh was a Senior Engineer at Hackerrank. In the last 5 years, he scaled the HackerRank community from 500k to 10M Developers.
+
+      He was also part of Navgurukul a Non Profit where he has worked on curriculum, jobs pipeline, and mentoring 100+ underprivileged students on software engineering.`,
+  },
+  // {
+  //   name: 'Naman Malhotra',
+  //   company: 'Co-founded Alconomy',
+  //   img: './mentor_nam_f.jpg',
+  //   linkedin: 'https://www.linkedin.com/in/naman03malhotra/',
+  //   descp: `Hi, I am Naman Malhotra. I still remember the first line of code I wrote when I was 13. Since then, I have worked with many startup founders across the continents.
+
+  //   I co-founded two startups; one of them was in the Ed-Tech space, and the other was in blockchain fintech. In both of them, I lead the entire tech end to end and product to some extent.
+
+  //   I strongly believe that if you have the right set of skills, your background, traditional education, etc hardly matters. Given that you are ready to work your ass off.`,
+  // },
+  {
+    name: 'Sai Ahladni Tripathy',
+    company: 'Ex - Amazon',
+    img: './mentor_sai.jpg',
+    linkedin: 'https://www.linkedin.com/in/ahladini/',
+    descp: `Sai was a software developer at Amazon. She is an Alumni of BITS Pilani.
+
+      She has previously worked as a placement co-ordinator at BITS and understands the recruitment market really well.
+
+      Apart from this Sai enjoys playing with her paw-friend.`,
+  },
+];
+
+const webinarData = [
+  /* {
+    name: 'Ankit Goyal',
+    title: 'Google',
+    descp: 'Graphs Minimum Spanning Trees and Interviewing at Google',
+    img: './mentors/Ankit_goyal_google.jpeg',
+  }, */
+
+  {
+    name: 'Rohit Aggarwal',
+    title: 'Lead Engineer Hackerrank',
+    descp: 'Building scalable backends and writing Microservices',
+    img: './mentors/rohit_hackerrank.jpeg',
+  },
+
+  {
+    name: 'Abhishek Gahlot',
+    title: 'Entrepreneur, EM at EduFlow',
+    descp: 'How browsers work, V8 engine and Networking in browsers',
+    img: './mentors/Abhishek_eduflow.jpeg',
+  },
+
+  {
+    name: 'Naman Malhotra',
+    title: 'Software Engineer at HackerRank',
+    descp: `Cracking Frontend Interviews and contributing to open source`,
+    img: './mentors/naman_hackerrank.jpg',
+  },
+
+  {
+    name: 'Vaibhav Gupta',
+    title: 'Nutanix Ex- Paypal',
+    descp: `System Designing and Creating a Highly Scalable URL Shortening service`,
+    img: './mentors/vaibhav_nutanix.jpeg',
+  },
+
+  {
+    name: 'Kritarth Anand',
+    title: 'Ex-Facebook',
+    descp: 'Importance of building a brand and having the right tribe',
+    img: './mentors/kritarth_facebook.jpg',
+  },
+
+  {
+    name: 'Kumar Akshay',
+    title: 'Atlassian',
+    descp: 'Starting with Opensource, GSOC and Interning at top-tech companies',
+    img: './mentors/akshay_atlassian.jpeg',
+  },
+
+  {
+    name: ' Sudhanshu Yadav',
+    title: 'Architect at HackerRank',
+    descp: 'Projects Review, JS V8 Engine and Contributing to Brahmos',
+    img: './mentors/sudhanshu_hackerrank.jpg',
+  },
+];
+
+export { prerequisite_data, who_can_join_data, mentorsData, webinarData };
